+++ src/nqdc/_entrez.py
@@ -1,9 +1,19 @@
 """Client for the Entrez E-utilities needed for downloading articles."""
+import io
 import logging
 import math
 import time
 from pathlib import Path
-from typing import Any, Dict, Mapping, Optional, Sequence, Union
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    Mapping,
+    Optional,
+    Sequence,
+    Tuple,
+    Union,
+)
 from urllib.parse import urljoin
 
 import requests
@@ -15,10 +25,72 @@ _LOG = logging.getLogger(__name__)
 _EFETCH_DEFAULT_BATCH_SIZE = 500
 
 
+def _check_response_status(response: requests.Response) -> Tuple[bool, str]:
+    """Check that a request was successful."""
+    if response.status_code == 200:
+        return True, ""
+    return False, f"Status code {response.status_code} != 200"
+
+
+def _check_efetch_response(response: requests.Response) -> Tuple[bool, str]:
+    """Check request was successful and content looks like an articleset."""
+    check, reason = _check_response_status(response)
+    if not check:
+        return check, reason
+    try:
+        parse_events = etree.iterparse(
+            io.BytesIO(response.content), events=("start",)
+        )
+        event, elem = next(parse_events)
+        assert event == "start"
+        assert elem.tag == "pmc-articleset"
+    except Exception:
+        return (
+            False,
+            "response content does not appear to be an XML articleset",
+        )
+    return True, ""
+
+
+def _check_esearch_response(response: requests.Response) -> Tuple[bool, str]:
+    """Check that esearch request was successful."""
+    check, reason = _check_response_status(response)
+    if not check:
+        return check, reason
+    try:
+        search_result: Dict[str, str] = response.json()["esearchresult"]
+    except Exception:
+        return False, "response does not contain JSON with esearchresult key"
+    if "ERROR" in search_result:
+        return False, f"esearch returned ERROR: {search_result['ERROR']}"
+    needed_keys = {"count", "webenv", "querykey"}
+    missing_keys = needed_keys.difference(search_result.keys())
+    if missing_keys:
+        return False, f"keys missing from esearch response: {missing_keys}"
+    return True, ""
+
+
+def _check_epost_response(response: requests.Response) -> Tuple[bool, str]:
+    check, reason = _check_response_status(response)
+    if not check:
+        return check, reason
+    try:
+        resp_xml = etree.fromstring(response.content)
+        webenv = resp_xml.find("WebEnv").text
+        assert webenv
+        querykey = resp_xml.find("QueryKey").text
+        assert querykey
+    except Exception:
+        return False, "epost response missing webenv or querykey"
+    return True, ""
+
+
 class EntrezClient:
     """Client for esearch and efetch using the pmc database."""
 
     _default_timeout = 27
+    _n_request_attempts = 5
+    _delay_before_retry_failed_request = 2.0
     _entrez_base_url = "https://eutils.ncbi.nlm.nih.gov/entrez/eutils/"
     _esearch_base_url = urljoin(_entrez_base_url, "esearch.fcgi")
     _efetch_base_url = urljoin(_entrez_base_url, "efetch.fcgi")
@@ -36,6 +108,8 @@ class EntrezClient:
             self._request_period = (
                 0.15 if "api_key" in self._entrez_id else 1.05
             )
+        else:
+            self._request_period = request_period
         self._last_request_time: Union[None, float] = None
         self._session = requests.Session()
         self.last_search_result: Optional[Mapping[str, str]] = None
@@ -51,27 +125,55 @@ class EntrezClient:
             time.sleep(wait)
         self._last_request_time = time.time()
 
-    def _send_request(
+    def _send_one_request(
         self,
-        url: str,
-        verb: str = "POST",
-        params: Optional[Mapping[str, Any]] = None,
-        data: Optional[Mapping[str, Any]] = None,
-    ) -> Union[None, requests.Response]:
-        req = requests.Request(verb, url, params=params, data=data)
-        prepped = self._session.prepare_request(req)
+        prepped: requests.PreparedRequest,
+        response_validator: Callable[[requests.Response], Tuple[bool, str]],
+    ) -> Optional[requests.Response]:
+        """Send prepared request and check response, return None if failed."""
         self._wait_to_send_request()
-        _LOG.debug(f"sending request: {prepped.url}")
         try:
             resp = self._session.send(prepped, timeout=self._default_timeout)
         except Exception:
-            _LOG.exception(f"Request failed: {url}")
+            _LOG.exception(f"Request failed: {prepped.url}")
+            return None
+        check, reason = response_validator(resp)
+        if not check:
+            _LOG.error(
+                f"Response failed to validate (reason: {reason}) "
+                f"for url {prepped.url}"
+            )
             return None
         _LOG.debug(
             f"received response. code: {resp.status_code}; "
             f"reason: {resp.reason}; from: {resp.url}"
         )
         return resp
+
+    def _send_request(
+        self,
+        url: str,
+        params: Optional[Mapping[str, Any]] = None,
+        data: Optional[Mapping[str, Any]] = None,
+        response_validator: Callable[
+            [requests.Response], Tuple[bool, str]
+        ] = _check_response_status,
+    ) -> Union[None, requests.Response]:
+        """Try to send a request several times.
+
+        Return the response when a request succeeds; return None if all
+        attempts fail.
+
+        """
+        req = requests.Request("POST", url, params=params, data=data)
+        prepped = self._session.prepare_request(req)
+        for attempt in range(self._n_request_attempts):
+            _LOG.debug(f"sending request: {prepped.url} (attempt #{attempt})")
+            resp = self._send_one_request(prepped, response_validator)
+            if resp is not None:
+                return resp
+            time.sleep(self._delay_before_retry_failed_request)
+        return None
 
     def epost(self, all_pmcids: Sequence[int]) -> Dict[str, str]:
         """Post a list of PMCIDs to the Entrez history server.
@@ -90,17 +192,17 @@ class EntrezClient:
             return {}
         params = {"db": "pmc", "id": ",".join(map(str, all_pmcids))}
         data = {**params, **self._entrez_id}
-        resp = self._send_request(self._epost_base_url, data=data, verb="POST")
+        resp = self._send_request(
+            self._epost_base_url,
+            data=data,
+            response_validator=_check_epost_response,
+        )
         if resp is None:
             self.n_failures = 1
             return {}
-        try:
-            resp_xml = etree.fromstring(resp.content)
-            webenv = resp_xml.find("WebEnv").text
-            querykey = resp_xml.find("QueryKey").text
-        except Exception:
-            self.n_failures = 1
-            return {}
+        resp_xml = etree.fromstring(resp.content)
+        webenv = resp_xml.find("WebEnv").text
+        querykey = resp_xml.find("QueryKey").text
         self.last_search_result = {
             "webenv": webenv,
             "querykey": querykey,
@@ -127,34 +229,29 @@ class EntrezClient:
         }
         data = {**search_params, **self._entrez_id}
         resp = self._send_request(
-            self._esearch_base_url, data=data, verb="POST"
+            self._esearch_base_url,
+            data=data,
+            response_validator=_check_esearch_response,
         )
         if resp is None:
             self.n_failures = 1
             return {}
-        try:
-            search_result: Dict[str, str] = resp.json()["esearchresult"]
-        except Exception:
-            self.n_failures = 1
-            return {}
-        if "ERROR" in search_result:
-            self.n_failures = 1
-            return {}
+        search_result: Dict[str, str] = resp.json()["esearchresult"]
         self.last_search_result = search_result
         _LOG.info(f"Search returned {search_result['count']} results")
         return search_result
 
-    def _check_search_result(
+    def _get_search_result(
         self,
         search_result: Optional[Mapping[str, str]] = None,
     ) -> Optional[Mapping[str, str]]:
+        """Return provided search result or result from last search.
+
+        If both are missing, return None.
+        """
         if search_result is None:
             search_result = self.last_search_result
         if search_result is None:
-            self.n_failures = 1
-            return None
-        needed_keys = {"count", "webenv", "querykey"}
-        if not needed_keys.issubset(search_result.keys()):
             self.n_failures = 1
             _LOG.error(
                 "Perform a search before calling `efetch`"
@@ -182,7 +279,7 @@ class EntrezClient:
         retmax are the same.
         """
         output_dir = Path(output_dir)
-        search_result = self._check_search_result(search_result)
+        search_result = self._get_search_result(search_result)
         if search_result is None:
             return
         search_count = int(search_result["count"])
@@ -222,9 +319,11 @@ class EntrezClient:
             return
         _LOG.info(f"getting batch {batch_nb + 1} / {n_batches}")
         resp = self._send_request(
-            self._efetch_base_url, verb="POST", data=params
+            self._efetch_base_url,
+            data=params,
+            response_validator=_check_efetch_response,
         )
-        if resp is None or resp.status_code != 200:
+        if resp is None:
             self.n_failures += 1
             _LOG.error(f"{self.n_failures} batches failed to download")
         else: